added insert/delete functions to SynchronizeAppsearchLocal
I added a function that handles inserted relationships and another one that handles deleted relationships.

I also added another operator that we might need and changed the operators used in the insert_person_relationship and delete_person_relationship. The relevant appsearch fields are defined as lists and only the relevant entries should be added/deleted.